[refactor] #106 NotificationStatus 엔티티 추가 및 Notification 내 상태 필드 제거
