+++ module-app/app-api/src/main/java/com/chaing/api/dto/notification/response/NotificationListResponse.java
@@ -10,16 +10,14 @@ public record NotificationListResponse(
         Long userId,
         NotificationType type,
         String message,
-        Long targetId,
-        boolean isRead
+        Long targetId
 ) {
     public static NotificationListResponse from(Notification notification) {
         return new NotificationListResponse(
                 notification.getUserId(),
                 notification.getType(),
                 notification.getMessage(),
-                notification.getTargetId(),
-                notification.isRead()
+                notification.getTargetId()
         );
     }
 }
