+++ module-domain/domain-notifications/src/main/java/com/chaing/domain/notifications/entity/NotificationStatus.java
@@ -1,0 +1,42 @@
+package com.chaing.domain.notifications.entity;
+
+import com.chaing.core.entity.BaseEntity;
+import jakarta.persistence.Column;
+import jakarta.persistence.Entity;
+import jakarta.persistence.GeneratedValue;
+import jakarta.persistence.GenerationType;
+import jakarta.persistence.Id;
+import lombok.AccessLevel;
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+@Entity
+@Getter
+@Builder
+@NoArgsConstructor(access = AccessLevel.PROTECTED)
+@AllArgsConstructor(access = AccessLevel.PROTECTED)
+public class NotificationStatus extends BaseEntity {
+
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private Long statusId;
+
+    @Column(nullable = false)
+    private Long userId;
+
+    @Column(nullable = false)
+    private Long notificationId;
+
+    @Column(nullable = false)
+    private boolean isRead;
+
+    public void read() {
+        this.isRead = true;
+    }
+
+    public void markAsUnread() {
+        this.isRead = false;
+    }
+}
